global Avatar refactorization
