+++ src/pages/level1/Level1.jsx
@@ -8,7 +8,7 @@ import Environments from "./staging/Environments";
 import { Canvas } from "@react-three/fiber";
 import Level1World from "./world/Level1World";
 import Controls from "./controls/Controls";
-import Avatar from "./characters/avatar/Avatar";
+import Avatar from "../../utils/avatar/Avatar";
 import useMovements from "../../utils/key-movements";
 import Ecctrl, { EcctrlAnimation } from "ecctrl";
 import Instructive from "../../utils/instructive/Instructive";
