+++ src/utils/avatar/Avatar.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from "react";
-import { useAvatar } from "../../../../context/AvatarContext";
+import { useAvatar } from "../../context/AvatarContext";
 import { useAnimations, useGLTF } from "@react-three/drei";
 
 export default function Avatar() {
